Put pg specific stuff in its own file
